# contracts/openapi/openapi.yaml
openapi: 3.0.0
info:
  title: Kanban API
  version: 1.1.0
...
paths:
...
  /api/columns/{columnId}/tasks:
    post:
      operationId: Tasks_create
      parameters:
        - name: columnId
          in: path
          required: true
          schema:
            type: string
      responses:
        '201':
          description: The request has succeeded and a new resource has been created as a result.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskItem'
        '400':
          description: The server could not understand the request due to invalid syntax.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      tags:
        - Tasks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaskRequest'
  /api/columns/{columnId}/tasks/{taskId}:
    patch:
      operationId: Tasks_update
      parameters:
        - name: columnId
          in: path
          required: true
          schema:
            type: string
        - name: taskId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskItem'
        '404':
          description: The server cannot find the requested resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      tags:
        - Tasks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTaskRequest'
    delete:
      operationId: Tasks_delete
      parameters:
        - name: columnId
          in: path
          required: true
          schema:
            type: string
        - name: taskId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: 'There is no content to send for this request, but the headers may be useful. '
        '404':
          description: The server cannot find the requested resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      tags:
        - Tasks
  /api/columns/{columnId}/tasks/{taskId}/move:
    patch:
      operationId: Tasks_move
      parameters:
        - name: columnId
          in: path
          required: true
          schema:
            type: string
        - name: taskId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskItem'
        '404':
          description: The server cannot find the requested resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      tags:
        - Tasks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MoveTaskRequest'
...
security:
  - {}
  - ApiKeyAuth: []
components:
  schemas:
...
    CreateTaskRequest:
      type: object
      required:
        - title
      properties:
        title:
          type: string
        description:
          type: string
        assigneeId:
          type: string
        priority:
          type: string
        tags:
          type: array
          items:
            type: string
        dueDate:
          type: string
          format: date-time
    ErrorResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: string
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
      description: Standard error response returned on failures
...
    MoveTaskRequest:
      type: object
      required:
        - targetColumnId
        - position
      properties:
        targetColumnId:
          type: string
        position:
          type: integer
          format: int32
...
    TaskItem:
      type: object
      required:
        - id
        - title
        - position
        - columnId
        - priority
        - tags
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        position:
          type: integer
          format: int32
        columnId:
          type: string
        assigneeId:
          type: string
        priority:
          type: string
        tags:
          type: array
          items:
            type: string
        dueDate:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    UpdateTaskRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        assigneeId:
          type: string
        priority:
          type: string
        tags:
          type: array
          items:
            type: string
        dueDate:
          type: string
          format: date-time
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: cookie
      name: session
servers:
  - url: http://localhost:5010
    description: Development server
    variables: {}